requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
argilla==1.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gptcache==0.1.11
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.4.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.141
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
olefile==0.46
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.1
pipreqs==0.4.13
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pypandoc==1.11
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
streamlit==1.21.0
streamlit-chat==0.0.2.2
tenacity==8.2.2
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unstructured==0.5.12
urllib3==1.26.15
validators==0.20.0
wrapt==1.14.1
XlsxWriter==3.1.0
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0
streamlit-talk
langchain
google-api-python-client
faiss-cpu
pypdf
pinecone-client
